stacks/iot-lab-bacnet/tb_bridge/requirements.txt
bacpypes==0.19.0
paho-mqtt==2.1.0
